dse: Move everything inside passes/clusters
