--- devito/dse/__init__.py
@@ -1,4 +1,0 @@
-from devito.dse.aliases import *  # noqa
-from devito.dse.manipulation import *  # noqa
-from devito.dse.rewriters import *  # noqa
-from devito.dse.transformer import *  # noqa
