+++ devito/passes/clusters/__init__.py
@@ -1,1 +1,5 @@
+from .aliases import *  # noqa
+from .manipulation import *  # noqa
+from .rewriters import *  # noqa
+from .transformer import *  # noqa
 from .misc import *  # noqa
