+++ devito/passes/clusters/manipulation.py
